dashboards/check-docs/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-frontmatter==1.0.1
pydantic==2.5.0
python-multipart==0.0.6